Documenation

+++ include/component_manager.hpp
@@ -1,3 +1,10 @@
+/**
+ * @file ComponentManager.hpp
+ * @brief Header file for the ComponentManager class.
+ *
+ * @defgroup EntityManager Entity Manager
+ * @brief This file defines the ComponentManager class, which manages the components attached to entities in the game world.
+ */
 #pragma once
 #include<unordered_map>
 #include<memory>
@@ -10,19 +17,98 @@
 #include "glm/gtc/matrix_transform.hpp"
 #include "glm/gtc/type_ptr.hpp"
 
+ /**
+	* @class ComponentManager
+	* @brief This class manages the components attached to entities in the game world.
+	*
+	* The ComponentManager class provides functionality to register component types,
+	*  create and access components for entities, and add/remove entities. It uses
+	*  an entity-component-system (ECS) approach for managing game objects.
+	*
+	* @ingroup EntityManager
+	*/
 struct ComponentManager {
 	std::unordered_map<std::size_t, std::unique_ptr<component_base>> component_classes_;
 	std::vector <size_t> deleted_components_;
 
+	/**
+	* Constructor for the ComponentManager class.
+	*
+	* Adds default components.
+	*/
 	ComponentManager();
+
+	/**
+	 * Destructor for the ComponentManager class.
+	 *
+	 */
 	~ComponentManager() = default;
 
+	/**
+	 * Template method to register a component class with the manager.
+	 *
+	 * This template method allows registering a specific component type with the manager.
+	 *  It internally creates a component list to store instances of that type.
+	 *
+	 * @tparam T The type of the component to register.
+	 */
 	template<typename T> void add_component_class();
-	template<typename T> void create_component(int position, T& component);
+
+	/**
+	* Template method to create a component of a specific type for an entity.
+	*
+	* This template method creates a component of the specified type and associates it with an entity.
+	*  It takes the entity ID and a reference to the component data as arguments.
+	*
+	* @tparam T The type of the component to create.
+	* @param position Index at which to insert the component in the entity's component list (optional).
+	* @param component Reference to the component data to create.
+	*/
+	template<typename T> void create_component(size_t position, T& component);
+
+	/**
+	 * Template method to retrieve a component of a specific type for an entity.
+	 *
+	 * This template method retrieves a component of the specified type for an entity identified by ID.
+	 *  It returns a pointer to the component data if it exists, or nullptr otherwise.
+	 *
+	 * @tparam T The type of the component to retrieve.
+	 * @param e The ID of the entity to retrieve the component from.
+	 * @return A pointer to the component data, or nullptr if not found.
+	 */
 	template<typename T> T* get_component(size_t e);
+
+	/**
+	 * Template method to retrieve a list of optional components of a specific type for all entities.
+	 *
+	 * This template method retrieves a reference to a vector containing std::optional objects for
+	 *  components of the specified type. Each element in the vector corresponds to an entity,
+	 *  and the optional object holds the component data if it exists for that entity.
+	 *
+	 * @tparam T The type of the component to retrieve.
+	 * @return A reference to the vector containing std::optional objects for the components.
+	 */
 	template<typename T> std::vector < std::optional<T>>* get_component_list();
 
+	/**
+	* Template method to create a new entity with optional components.
+	*
+	* This template method creates a new entity and optionally adds components of the specified types.
+  * It takes a variadic template parameter pack to represent the component types.
+	* 
+	* @tparam ... Ts Additional component types to add to the entity (variadic template parameter pack).
+	* @return The ID of the newly created entity.
+	*/
 	template<class ... T> size_t add_entity(T... components);
+
+	/**
+	 * Method to remove an entity from the manager.
+	 *
+	 * This method removes an entity identified by ID from the manager. It marks the entity's
+	 *  components for potential reuse and updates internal data structures.
+	 *
+	 * @param id The ID of the entity to remove.
+	 */
 	void remove_entity(size_t id);
 
 };
@@ -44,7 +130,7 @@ template<typename T> T* ComponentManager::get_component(size_t e) {
 	return &component_opt.value();
 }
 
-template<typename T> void ComponentManager::create_component(int position, T& component) {
+template<typename T> void ComponentManager::create_component(size_t position, T& component) {
 	//Cast to the specific component list and add the components
 	component_list<T>* list = dynamic_cast<component_list<T>*>(component_classes_[typeid(T).hash_code()].get());
 	list->add_component_at(position, component);
@@ -59,7 +145,7 @@ template<typename T> std::vector < std::optional<T>>* ComponentManager::get_comp
 }
 
 template<class ... T> size_t ComponentManager::add_entity(T... components){
-	if (deleted_components_.size() > 0) {
+	if ((int)(deleted_components_.size()) > 0) {
 		size_t id = deleted_components_.back();
 		deleted_components_.pop_back();
 
@@ -68,9 +154,9 @@ template<class ... T> size_t ComponentManager::add_entity(T... components){
 		}
 		return id;
 	}
-
+	
 	size_t size = 0;
-
+	
 	//Add nullopt to components that not matches with the parameters
 	for (auto& [key, value] : component_classes_) {
 		value->add_component();
@@ -79,7 +165,7 @@ template<class ... T> size_t ComponentManager::add_entity(T... components){
 
 	//Fold expression to iterate in parameters pack (components)
 	(create_component(size, components), ...);
-
+	
 	//Return entity value
 	return size;
 }
